implemented the new test loop for ease of use bluetooth functionality for remote testing
