+++ goomba_driving.py
@@ -23,14 +23,13 @@ from math import cos
 from math import sin
 
 from goomba_state import state
-''' probably unnecessary bluetooth stuff
-from adafruit_ble import BLERadio
+
+from adafruit_ble import BLERadio  # for testing motors remotely
 from adafruit_ble.advertising.standard import ProvideServicesAdvertisement
 from adafruit_ble.services.nordic import UARTService
 
 from adafruit_bluefruit_connect.packet import Packet
 from adafruit_bluefruit_connect.button_packet import ButtonPacket
-'''
 
 '''
  we have 7 digital pins, 6 analog pins explicitly - D13, D12, D11, D10, D9, D6, D5, D2
@@ -103,45 +102,22 @@ raspi = SPIDevice(spi, cs, baudrate=5000000, polarity=0, phase=0)  # need to loo
 '''
 
 DUTY_MAX = 2**16-1
-blink_time = 0.1
-last_time = 0
+
+ble = BLERadio()
+uart = UARTService()
+advertisement = ProvideServicesAdvertisement(uart)
+
+B1 = ButtonPacket.BUTTON_1
+B2 = ButtonPacket.BUTTON_2
+B3 = ButtonPacket.BUTTON_3
+B4 = ButtonPacket.BUTTON_4
+DOWN = ButtonPacket.DOWN
+UP = ButtonPacket.UP
+LEFT = ButtonPacket.LEFT
+RIGHT = ButtonPacket.RIGHT
+
 
 # States of state machine
-class forward_state(state):
-
-    def on_event(self, event, dist, ang):
-        if event == "FRONT":
-            # some way to output distance of wall here
-            return ["TURN", "LEFT"]  # i want to check for multiple events at once?
-            # or could i just say turn on direction until there is no longer a front event
-            # also want something in here to store distance of object
-
-        elif event == "LEFT":
-            # some way to output distance of wall here
-            stuff = 0
-        elif event == "RIGHT":
-            # some way to output distance of wall here
-            stuff = 0
-        elif event == "CLIFF":
-            stuff = 0
-        else:
-            return self
-
-
-class turn_state(state):
-
-    def on_event(self):
-        again = 1
-        noty = 2
-        stuff_here = 1
-        testing_stuff = again + noty * stuff_here
-
-
-class idling_state(state):
-    def on_event(self):
-        idk = again
-
-
 class state_machine(object):
     def on_event(self, initial_state):
         self.state = idling_state()
@@ -222,7 +198,18 @@ def vector_store(vec_list, outfile):
     # JSON stuff here?
     json = "jason"
 
-print(0)
+def motor_test(mot1, mot2, drive_time):
+    drive = drive_time/3
+    motor_level(60, mot1)
+    motor_level(60, mot2)
+    time.sleep(drive)
+    motor_level(-60, mot1)
+    motor_level(-60, mot2)
+    time.sleep(drive)
+    motor_level(0, mot1)
+    motor_level(0, mot2)
+    time.sleep(drive)
+
 # how to use SPI
 with raspi:
     result = bytearray(4)  # 4 byte buffer is created to hold the result of the SPI read
@@ -239,9 +226,7 @@ with raspi:
 '''
 
 # Main loop
-vector_array = {}
-origins = [[(0, 0, 0), (0, 0)], [(0, 0, 0), (0, 0)]]  # would this part be easier with object oriented programming?
-i = 0
+
 # to try multiple I2C devices for sunday
 #while not i2c.try_lock():
     #pass
@@ -254,19 +239,76 @@ except:
 # Unlock I2C now that we're done scanning.
 i2c.unlock()
 # https://e2e.ti.com/blogs_/b/analogwire/posts/how-to-simplify-i2c-tree-when-connecting-multiple-slaves-to-an-i2c-master
-event = "IDLE"
-while True:
-    if i == 0:
-        origins[i][1] = (0, 0)  # to store vector w/ magnetic data
-    if event == "TURN":
-        i += 1
-        origins[i][0] = lis3.magnetic
-    mag_x, mag_y, mag_z = origins[i][0]
-    dists = [sonarL.distance, sonarF.distance, sonarR.distance]
-    distL, distF, distR = dists
 
-    print("Sonar distances: {:.2f}L {:.2f}F {:.2f}R (cm)".format(*dists))
-    print('Magnetometer: {0:10.2f}X {1:10.2f}Y {2:10.2f}Z uT'.format(mag_x, mag_y, mag_z))
-    print('Encoders: {0:10.2f}L {1:10.2f}R pulses'.format(encL.position, encR.position))
 
-    time.sleep(0.25)
+step = 5
+speed1 = 0
+speed2 = 0
+while True:  # the testing loop
+    ble.start_advertising(advertisement)
+    while not ble.connected:
+        dists = [sonarL.distance, sonarF.distance, sonarR.distance]
+
+        print("Sonar distances: {:.2f}L {:.2f}F {:.2f}R (cm)".format(*dists))
+        print('Magnetometer: {0:10.2f}X {1:10.2f}Y {2:10.2f}Z uT'.format(*lis3.magnetic))
+        print('Encoders: {0:10.2f}L {1:10.2f}R pulses'.format(encL.position, encR.position))
+
+        time.sleep(0.5)
+        mot_test0 = input("Test motors? /n Y or N ")
+        mot_test1 = mot_test0.upper
+        if mot_test1 == "Y":
+            duration = float(input("How long? "))
+            motor_test(motL, motR, duration)
+        elif mot_test1 == "END":
+            break
+
+    # Now we're connected
+
+    while ble.connected:
+        if uart.in_waiting:
+            packet = Packet.from_stream(uart)
+            if isinstance(packet, ButtonPacket):
+                if packet.pressed:
+                    if packet.button == B1:
+                        # The 1 button was pressed.
+                        print("1 button pressed! It was not very effective.")
+                    elif packet.button == UP:
+                        # The UP button was pressed.
+                        print("UP button pressed! The left motor's speed sharply rose!")
+                        if speed1 != 100:
+                            speed1 += step
+                    elif packet.button == DOWN:
+                        # The DOWN button was pressed.
+                        print("DOWN button pressed! The  left motor's speed sharply fell!")
+                        if speed1 != -100:
+                            speed1 += -step
+                    elif packet.button == LEFT:
+                        # The LEFT button was pressed.
+                        print("LEFT button pressed! The right motor's speed sharply fell!")
+                        if speed2 != -100:
+                            speed2 += -step
+                    elif packet.button == RIGHT:
+                        # The RIGHT button was pressed.
+                        print("RIGHT button pressed! The right motor's speed sharply rose!")
+                        if speed2 != 100:
+                            speed2 += step
+                    elif packet.button == B2:
+                        # The 2 button was pressed.
+                        print("2 button pressed! It was not very effective.")
+                    elif packet.button == B3:
+                        # The 3 button was pressed.
+                        print("3 button pressed! It was not very effective.")
+                    elif packet.button == B4:
+                        # The 4 button was pressed.
+                        print("4 button pressed! It was not very effective.")
+        
+        motor_level(speed1, motL)
+        motor_level(speed1, motR)
+
+        dists = [sonarL.distance, sonarF.distance, sonarR.distance]
+
+        print("Sonar distances: {:.2f}L {:.2f}F {:.2f}R (cm)".format(*dists))
+        print('Magnetometer: {0:10.2f}X {1:10.2f}Y {2:10.2f}Z uT'.format(*lis3.magnetic))
+        print('Encoders: {0:10.2f}L {1:10.2f}R pulses'.format(encL.position, encR.position))
+
+        time.sleep(0.1)
